Added pipeline settings

+++ Jenkinsfile
@@ -5,11 +5,22 @@ pipeline {
         registry = "197.26.204.208:8083"
     }
     stages {
-
+    stage('Clean Build Folder (Optional)') {
+            steps {
+                script {
+                    if (!fileExists('build')) {
+                        echo 'Build folder not found.'
+                    } else {
+                        sh 'rm -rf build'
+                        sh 'ls'
+                        echo 'Build folder removed.'
+                    }
+                }
+            }
+        }
         stage('Build application') {
             steps {
                 script {
-                        sh 'rm -rf build'
                         sh 'npm run build'
                 }
             }
